[UPDATE] Video Update api 수정, id Long type으로 통일

+++ src/main/java/com/semtleWebGroup/youtubeclone/domain/video/repository/VideoInfoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface VideoInfoRepository extends JpaRepository<VideoInfo, Integer> {
+public interface VideoInfoRepository extends JpaRepository<VideoInfo, Long> {
     List<VideoInfo> findByTitleContaining(String key);
